Refactor ZipObject to ZipObjectReader
