+++ src/inc/ObjectBase.hpp
@@ -9,6 +9,7 @@
 #include <type_traits>
 #include "Exceptions.hpp"
 #include "StreamBase.hpp"
+#include "MsixFeatureSelector.hpp"
 
 namespace MSIX { namespace Meta {
 //////////////////////////////////////////////////////////////////////////////////////////////
@@ -46,20 +47,69 @@ public:
     FieldBase() = default;
 
     size_t Size() { return sizeof(T); }
+    std::vector<std::uint8_t> GetBytes() { return std::vector<std::uint8_t>(); }
     
     T value;
 };
 
 // Simple 2, 4, and 8 byte fields
-class Field2Bytes final : public FieldBase<std::uint16_t> { };
-class Field4Bytes final : public FieldBase<std::uint32_t> { };
-class Field8Bytes final : public FieldBase<std::uint64_t> { };
+class Field2Bytes final : public FieldBase<std::uint16_t>
+{
+public:
+    std::vector<std::uint8_t> GetBytes()
+    {
+        THROW_IF_PACK_NOT_ENABLED
+        std::vector<std::uint8_t> result;
+        result.push_back(static_cast<std::uint8_t>(this->value));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 8));
+        return result;
+    }
+};
+
+class Field4Bytes final : public FieldBase<std::uint32_t>
+{
+public:
+    std::vector<std::uint8_t> GetBytes()
+    {
+        THROW_IF_PACK_NOT_ENABLED
+        std::vector<std::uint8_t> result;
+        result.push_back(static_cast<std::uint8_t>(this->value));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 8));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 16));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 24));
+        return result;
+    }
+};
+
+class Field8Bytes final : public FieldBase<std::uint64_t>
+{
+public:
+    std::vector<std::uint8_t> GetBytes()
+    {
+        THROW_IF_PACK_NOT_ENABLED
+        std::vector<std::uint8_t> result;
+        result.push_back(static_cast<std::uint8_t>(this->value));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 8));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 16));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 24));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 32));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 40));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 48));
+        result.push_back(static_cast<std::uint8_t>(this->value >> 56));
+        return result;
+    }    
+};
 
 // variable length field.
 class FieldNBytes : public FieldBase<std::vector<std::uint8_t>>
 {
 public:
     size_t Size() { return this->value.size(); }
+    std::vector<std::uint8_t> GetBytes()
+    {
+        THROW_IF_PACK_NOT_ENABLED
+        return this->value;
+    }
 };
 
 //////////////////////////////////////////////////////////////////////////////////////////////
@@ -99,6 +149,26 @@ public:
         {   result += field.Size();
         }, result);
         return result;
+    }
+
+    std::vector<std::uint8_t> GetBytes()
+    {
+        THROW_IF_PACK_NOT_ENABLED
+        std::vector<std::uint8_t> bytes;
+        this->for_each([](auto& field, std::size_t index, std::vector<std::uint8_t>& bytes)
+        {
+            auto fieldBytes = field.GetBytes();
+            bytes.insert(bytes.end(), fieldBytes.begin(), fieldBytes.end());
+        }, bytes);
+        return bytes;
+    }
+
+    void Write(const ComPtr<IStream>& stream)
+    {
+        THROW_IF_PACK_NOT_ENABLED
+        auto bytes = GetBytes();
+        ULONG bytesWritten = 0;
+        ThrowHrIfFailed(stream->Write(bytes.data(), static_cast<ULONG>(bytes.size()), &bytesWritten));
     }
 };
