+++ test/Win32/Win32.ps1
@@ -142,7 +142,7 @@ RunTest 0x8bad0042 .\..\appx\SignedTamperedContentTypes-TRUST_E_BAD_DIGEST.appx
 RunTest 0x8bad0042 .\..\appx\SignedUntrustedCert-CERT_E_CHAINING.appx
 RunTest 0x00000000 .\..\appx\TestAppxPackage_Win32.appx "-ss"
 RunTest 0x00000000 .\..\appx\TestAppxPackage_x64.appx "-ss"
-RunTest 0x8bad0012 .\..\appx\UnsignedZip64WithCI-APPX_E_MISSING_REQUIRED_FILE.appx
+RunTest 0x8bad0031 .\..\appx\UnsignedZip64WithCI-APPX_E_MISSING_REQUIRED_FILE.appx
 RunTest 0x8bad0001 .\..\appx\FileDoesNotExist.appx "-ss"
 RunTest 0x8bad0051 .\..\appx\BlockMap\Missing_Manifest_in_blockmap.appx "-ss"
 RunTest 0x8bad0051 .\..\appx\BlockMap\ContentTypes_in_blockmap.appx "-ss"
